# static/openapi.yaml
openapi: 3.1.0
info:
  title: CRM REST API
  version: 1.0.0
  summary: Документация REST API CRM
  description: |
    Описание REST API для CRM-приложения. Большинство эндпоинтов требует
    авторизации через cookie-сессию и защиты CSRF для методов изменения данных.

    В проекте используется cookie-сессия Flask-Login и CSRF-защита Flask-WTF.
    Для POST/PUT/DELETE запросов добавляйте заголовок `X-CSRFToken` с актуальным токеном.
    Получить новый токен можно через `GET /refresh_csrf`.
servers:
  - url: /
    description: Текущий сервер
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie-сессия Flask-Login (`session`)
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRFToken
      description: CSRF-токен Flask-WTF. Получайте через GET /refresh_csrf
  parameters:
    CsrfHeader:
      name: X-CSRFToken
      in: header
      required: true
      description: CSRF-токен для защищённых операций
      schema:
        type: string
...
  schemas:
...
    SearchResultItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        similarity:
          type: number
          format: float
      required: [id, name]
...
security:
  - cookieAuth: []
paths:
...
  /search/search_objects:
    get:
      tags: [Search]
      summary: Поиск объектов (подсказки)
      description: Возвращает до 10 объектов по подстроке.
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResultItem'
        '400': { description: Ошибка валидации запроса }
        '500': { description: Ошибка сервера }
  /search/search_contractors:
    get:
      tags: [Search]
      summary: Поиск подрядчиков (подсказки)
      description: Возвращает до 10 подрядчиков по подстроке.
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResultItem'
        '400': { description: Ошибка валидации запроса }
        '500': { description: Ошибка сервера }
  /search/search:
    get:
      tags: [Search]
      summary: Поиск (объект/подрядчик)
      description: Выполняет общий поиск по типу и запросу.
      parameters:
        - name: query
          in: query
          required: true
          schema: { type: string }
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [object, contractor]
      responses:
        '200':
          description: Результаты поиска
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchResultItem'
                  query: { type: string }
                required: [items, query]
        '500': { description: Ошибка сервера }
...
  /api/v1/audit/event:
    post:
      tags: [Audit]
      summary: Клиентское событие
      description: Приём клиентских событий (sendBeacon/JSON). Требуется CSRF-токен; при задании `AUDIT_EVENT_TOKEN` добавьте заголовок `X-Audit-Token`.
      security:
        - cookieAuth: []
          csrfHeader: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                data: { type: object }
            examples:
              click:
                value:
                  name: 'ui.click'
                  data: { button: 'save', section: 'requests' }
      responses:
        '204': { description: Принято }